+++ biopandas/mmtf/pandas_mmtf.py
@@ -1,11 +1,11 @@
 """Class for working with MMTF files."""
+
 from __future__ import annotations
 
 import os
 import copy
 import gzip
 import warnings
-from string import ascii_uppercase
 from typing import Any, Dict, List, Union
 from warnings import warn
 
@@ -58,12 +58,7 @@ class PandasMmtf(object):
         return self
 
     def fetch_mmtf(self, pdb_code: str):
-        self.code = pdb_code
-        self.mmtf = fetch(pdb_code)
-        df = self._mmtf_to_df(self.mmtf)
-        self._df["ATOM"] = df.loc[df.record_name == "ATOM"]
-        self._df["HETATM"] = df.loc[df.record_name == "HETATM"]
-        return self
+        raise DeprecationWarning("PDB No longer serves MMTF files.")
 
     @staticmethod
     def _mmtf_to_df(mmtf_obj: MMTFDecoder) -> pd.DataFrame:
@@ -95,9 +90,9 @@ class PandasMmtf(object):
                 t[d] = self.df[d].copy()
 
         for sec in records:
-            t[sec]["element_symbol"] = t[sec][["atom_name", "element_symbol"]].apply(
-                lambda x: x[0][1] if len(x[1]) == 3 else x[0][0], axis=1
-            )
+            t[sec]["element_symbol"] = t[sec][
+                ["atom_name", "element_symbol"]
+            ].apply(lambda x: x[0][1] if len(x[1]) == 3 else x[0][0], axis=1)
         return t
 
     @staticmethod
@@ -240,14 +235,18 @@ class PandasMmtf(object):
         cmp = "placeholder"
         indices = []
 
-        residue_number_insertion = tmp["residue_number"].astype(str) + tmp["insertion"]
+        residue_number_insertion = (
+            tmp["residue_number"].astype(str) + tmp["insertion"]
+        )
 
         for num, ind in zip(residue_number_insertion, np.arange(tmp.shape[0])):
             if num != cmp:
                 indices.append(ind)
             cmp = num
 
-        transl = tmp.iloc[indices][residue_col].map(amino3to1dict).fillna(fillna)
+        transl = (
+            tmp.iloc[indices][residue_col].map(amino3to1dict).fillna(fillna)
+        )
 
         return pd.concat((tmp.iloc[indices]["chain_id"], transl), axis=1)
 
@@ -287,7 +286,9 @@ class PandasMmtf(object):
 
         return np.sqrt(
             np.sum(
-                df[["x_coord", "y_coord", "z_coord"]].subtract(xyz, axis=1) ** 2, axis=1
+                df[["x_coord", "y_coord", "z_coord"]].subtract(xyz, axis=1)
+                ** 2,
+                axis=1,
             )
         )
 
@@ -313,7 +314,9 @@ class PandasMmtf(object):
         """
         return np.sqrt(
             np.sum(
-                df[["x_coord", "y_coord", "z_coord"]].subtract(xyz, axis=1) ** 2, axis=1
+                df[["x_coord", "y_coord", "z_coord"]].subtract(xyz, axis=1)
+                ** 2,
+                axis=1,
             )
         )
 
@@ -359,7 +362,9 @@ class PandasMmtf(object):
                 if c in {"x_coord", "y_coord", "z_coord"}:
                     for idx in range(dfs[r][c].values.shape[0]):
                         if len(dfs[r][c].values[idx]) > 8:
-                            dfs[r][c].values[idx] = str(dfs[r][c].values[idx]).strip()
+                            dfs[r][c].values[idx] = str(
+                                dfs[r][c].values[idx]
+                            ).strip()
                 if c in {"line_idx", "OUT"}:
                     pass
                 elif r in {"ATOM", "HETATM"} and c not in pdb_df_columns:
@@ -403,7 +408,7 @@ class PandasMmtf(object):
         """Parse secondary structure elements"""
         raise NotImplementedError
 
-    def to_mmtf(self, path, records = ("ATOM", "HETATM")):
+    def to_mmtf(self, path, records=("ATOM", "HETATM")):
         """Write record DataFrames to an MMTF file.
 
         Parameters
@@ -417,7 +422,6 @@ class PandasMmtf(object):
         """
         df = pd.concat(objs=[self.df[i] for i in records])
         return write_mmtf(df, path)
-
 
     def get_model(self, model_index: int) -> PandasMmtf:
         """Returns a new PandasPDB object with the dataframes subset to the
@@ -434,19 +438,21 @@ class PandasMmtf(object):
             structure subsetted to the given model.
         """
 
-        df = copy.deepcopy(self)
+        biopandas_structure = copy.deepcopy(self)
 
-        if "ATOM" in df.df.keys():
-            df.df["ATOM"] = df.df["ATOM"].loc[df.df["ATOM"]["model_id"] == model_index]
-        if "HETATM" in df.df.keys():
-            df.df["HETATM"] = df.df["HETATM"].loc[
-                df.df["HETATM"]["model_id"] == model_index
+        if "ATOM" in biopandas_structure.df.keys():
+            biopandas_structure.df["ATOM"] = biopandas_structure.df["ATOM"].loc[
+                biopandas_structure.df["ATOM"]["model_id"] == model_index
             ]
-        if "ANISOU" in df.df.keys():
-            df.df["ANISOU"] = df.df["ANISOU"].loc[
-                df.df["ANISOU"]["model_id"] == model_index
+        if "HETATM" in biopandas_structure.df.keys():
+            biopandas_structure.df["HETATM"] = biopandas_structure.df["HETATM"].loc[
+                biopandas_structure.df["HETATM"]["model_id"] == model_index
             ]
-        return df
+        if "ANISOU" in biopandas_structure.df.keys():
+            biopandas_structure.df["ANISOU"] = biopandas_structure.df["ANISOU"].loc[
+                biopandas_structure.df["ANISOU"]["model_id"] == model_index
+            ]
+        return biopandas_structure
 
     def get_models(self, model_indices: List[int]) -> PandasMmtf:
         """Returns a new PandasMmtf object with the dataframes subset to the
@@ -463,22 +469,30 @@ class PandasMmtf(object):
             containing the structure subsetted to the given model.
         """
 
-        df = copy.deepcopy(self)
+        biopandas_structure = copy.deepcopy(self)
 
-        if "ATOM" in df.df.keys():
-            df.df["ATOM"] = df.df["ATOM"].loc[
-                [x in model_indices for x in df.df["ATOM"]["model_id"].tolist()]
+        if "ATOM" in biopandas_structure.df.keys():
+            biopandas_structure.df["ATOM"] = biopandas_structure.df["ATOM"].loc[
+                [
+                    x in model_indices
+                    for x in biopandas_structure.df["ATOM"]["model_id"].tolist()
+                ]
             ]
-        if "HETATM" in df.df.keys():
-            df.df["HETATM"] = df.df["HETATM"].loc[
-                [x in model_indices for x in df.df["HETATM"]["model_id"].tolist()]
+        if "HETATM" in biopandas_structure.df.keys():
+            biopandas_structure.df["HETATM"] = biopandas_structure.df["HETATM"].loc[
+                [
+                    x in model_indices
+                    for x in biopandas_structure.df["HETATM"]["model_id"].tolist()
+                ]
             ]
-        if "ANISOU" in df.df.keys():
-            df.df["ANISOU"] = df.df["ANISOU"].loc[
-                [x in model_indices for x in df.df["ANISOU"]["model_id"].tolist()]
+        if "ANISOU" in biopandas_structure.df.keys():
+            biopandas_structure.df["ANISOU"] = biopandas_structure.df["ANISOU"].loc[
+                [
+                    x in model_indices
+                    for x in biopandas_structure.df["ANISOU"]["model_id"].tolist()
+                ]
             ]
-        return df
-
+        return biopandas_structure
 
 
 def fetch_mmtf(pdb_code: str) -> pd.DataFrame:
@@ -501,7 +515,11 @@ def parse_mmtf(file_path: str) -> pd.DataFrame:
     :return: Dataframe of protein structure.
     :rtype: pd.DataFrame
     """
-    df = parse_gzip(file_path) if file_path.endswith(".gz") else parse(file_path)
+    df = (
+        parse_gzip(file_path)
+        if file_path.endswith(".gz")
+        else parse(file_path)
+    )
     return mmtf_to_df(df)
 
 
@@ -533,7 +551,9 @@ def mmtf_to_df(mmtf_obj: MMTFDecoder) -> pd.DataFrame:
         else:
             chain_indices[i] = chain_indices[i] + chain_indices[i - 1]
     model_indices = mmtf_obj.chains_per_model
-    model_indices = [sum(model_indices[:i+1]) for i in range(len(model_indices))]
+    model_indices = [
+        sum(model_indices[: i + 1]) for i in range(len(model_indices))
+    ]
     ch_idx = 0
 
     entity_types = {}
@@ -541,23 +561,29 @@ def mmtf_to_df(mmtf_obj: MMTFDecoder) -> pd.DataFrame:
         for chain in i["chainIndexList"]:
             entity_types[chain] = i["type"]
 
+    ch_idx_iter = iter(sorted(list(entity_types.keys())))
+    ch_idx = next(ch_idx_iter)
     for idx, i in enumerate(mmtf_obj.group_type_list):
         res = mmtf_obj.group_list[i]
-        #record = "HETATM" if res["chemCompType"] == "NON-POLYMER" else "ATOM"
-        #record = (
+        # record = "HETATM" if res["chemCompType"] == "NON-POLYMER" else "ATOM"
+        # record = (
         #    "ATOM"
         #    if res["chemCompType"] in ["L-PEPTIDE LINKING", "PEPTIDE LINKING"]
         #    else "HETATM"
-        #)
+        # )
         if idx == chain_indices[ch_idx]:
-            ch_idx += 1
+            # ch_idx += 1
+            ch_idx = next(ch_idx_iter)
         record = "ATOM" if entity_types[ch_idx] == "polymer" else "HETATM"
 
         for _ in res["atomNameList"]:
             data["residue_name"].append(res["groupName"])
             data["residue_number"].append(mmtf_obj.group_id_list[idx])
+            # data["chain_id"].append([mmtf_obj.chain_name_list[ch_idx]])
             data["chain_id"].append([mmtf_obj.chain_name_list[ch_idx]])
-            data["model_id"].append(int(np.argwhere(np.array(model_indices)>ch_idx)[0]) + 1)
+            data["model_id"].append(
+                int(np.argwhere(np.array(model_indices) > ch_idx)[0]) + 1
+            )
             data["record_name"].append(record)
             data["insertion"].append(mmtf_obj.ins_code_list[idx])
         data["atom_name"].append(res["atomNameList"])
@@ -577,15 +603,18 @@ def mmtf_to_df(mmtf_obj: MMTFDecoder) -> pd.DataFrame:
             "record_name",
             "insertion",
             "atom_number",
-            "model_id"
+            "model_id",
         ]:
             continue
         data[k] = [i for sublist in v for i in sublist]
 
-    df = pd.DataFrame.from_dict(data).sort_values(by=["model_id", "atom_number"])
+    df = pd.DataFrame.from_dict(data).sort_values(
+        by=["model_id", "atom_number"]
+    )
     df.alt_loc = df.alt_loc.str.replace("\x00", "")
     df.insertion = df.insertion.str.replace("\x00", "")
     return df
+
 
 def _seq1(seq, charmap: Dict[str, str], undef_code="X"):
     # sourcery skip: dict-assign-update-to-union
@@ -626,10 +655,8 @@ def _seq1(seq, charmap: Dict[str, str], undef_code="X"):
     onecode = {k.upper(): v for k, v in charmap.items()}
     # add the given termination codon code and custom maps
     onecode.update((k.upper(), v) for k, v in charmap.items())
-    seqlist = [seq[3 * i : 3 * (i + 1)] for i in range(len(seq) // 3)]
+    seqlist = [seq[3 * i:3 * (i + 1)] for i in range(len(seq) // 3)]
     return "".join(onecode.get(aa.upper(), undef_code) for aa in seqlist)
-
-
 
 
 def write_mmtf(df: pd.DataFrame, file_path: str):
@@ -669,7 +696,17 @@ def write_mmtf(df: pd.DataFrame, file_path: str):
         experimental_methods=None,
     )
 
-    node_ids = df.model_id.astype(str) + ":" + df.chain_id + ":" + df.residue_name + ":" + df.residue_number.astype(str) + ":" + df.insertion.astype(str)
+    node_ids = (
+        df.model_id.astype(str)
+        + ":"
+        + df.chain_id
+        + ":"
+        + df.residue_name
+        + ":"
+        + df.residue_number.astype(str)
+        + ":"
+        + df.insertion.astype(str)
+    )
     df["residue_id"] = node_ids
     # Tracks values to replace them at the end
     chains_per_model = []
@@ -685,10 +722,10 @@ def write_mmtf(df: pd.DataFrame, file_path: str):
         count_models += 1
         # Set the model info
         encoder.set_model_info(
-            #model_id=model_idx, # According to mmtf-python this is meaningless
-            model_id=model_idx, # According to mmtf-python this is meaningless
-            chain_count=0 # Set to 0 here and changed later
-            )
+            # model_id=model_idx, # According to mmtf-python this is meaningless
+            model_id=model_idx,  # According to mmtf-python this is meaningless
+            chain_count=0,  # Set to 0 here and changed later
+        )
         # Iterate over chains in model
         for chain_id in chains:
             seqs = []
@@ -726,51 +763,73 @@ def write_mmtf(df: pd.DataFrame, file_path: str):
                 #  structure object so we treat each molecule as a separate
                 #  entity
                 if residue_type != prev_res_type or (
-                        residue_type == "HETATM" and resname != prev_resname
-                    ):
+                    residue_type == "HETATM" and resname != prev_resname
+                ):
                     encoder.set_entity_info(
                         chain_indices=[count_chains],
-                        sequence="", # Set to empty here and changed later
+                        sequence="",  # Set to empty here and changed later
                         description="",
                         entity_type=entity_type,
                     )
                     encoder.set_chain_info(
                         chain_id=chain_id,
-                        chain_name="\x00" if len(chain_id.strip()) == 0 else chain_id,
-                        num_groups=0, # Set to 0 here and changed later
+                        chain_name=(
+                            "\x00" if len(chain_id.strip()) == 0 else chain_id
+                        ),
+                        num_groups=0,  # Set to 0 here and changed later
                     )
                     if count_chains > 0:
-                        groups_per_chain.append(count_groups - sum(groups_per_chain) -1)
+                        groups_per_chain.append(
+                            count_groups - sum(groups_per_chain) - 1
+                        )
                     if not first_chain:
                         seqs.append(seq)
                     first_chain = False
                     count_chains += 1
-                    seq=""
+                    seq = ""
 
                 if entity_type == "polymer":
-                    seq += _seq1(residue_df.residue_name.unique()[0], charmap=protein_letters_3to1_extended)
+                    seq += _seq1(
+                        residue_df.residue_name.unique()[0],
+                        charmap=protein_letters_3to1_extended,
+                    )
 
                 prev_res_type = residue_type
                 prev_resname = resname
 
-                group_type = "NON-POLYMER" if residue_type == "HETATM" else "L-PEPTIDE LINKING"
+                group_type = (
+                    "NON-POLYMER"
+                    if residue_type == "HETATM"
+                    else "L-PEPTIDE LINKING"
+                )
                 encoder.set_group_info(
                     group_name=residue_df.residue_name.unique()[0],
                     group_number=int(residue_df.residue_number.unique()[0]),
-                    insertion_code="\x00" if residue_df.insertion.unique()[0] == "" else residue_df.insertion.unique()[0],
-                    group_type=group_type, # Hack to ensure we can re-parse.
+                    insertion_code=(
+                        "\x00"
+                        if residue_df.insertion.unique()[0] == ""
+                        else residue_df.insertion.unique()[0]
+                    ),
+                    group_type=group_type,  # Hack to ensure we can re-parse.
                     atom_count=len(residue_df),
                     bond_count=0,
-                    single_letter_code=_seq1(df.residue_name.unique()[0], charmap=protein_letters_3to1_extended),
-                    sequence_index=len(seq) - 1 if entity_type == "polymer" else -1,
-                    secondary_structure_type=-1
+                    single_letter_code=_seq1(
+                        df.residue_name.unique()[0],
+                        charmap=protein_letters_3to1_extended,
+                    ),
+                    sequence_index=(
+                        len(seq) - 1 if entity_type == "polymer" else -1
+                    ),
+                    secondary_structure_type=-1,
                 )
                 for row in residue_df.itertuples():
                     count_atoms += 1
                     encoder.set_atom_info(
                         atom_name=row.atom_name,
                         serial_number=row.atom_number,
-                        alternative_location_id="\x00" if row.alt_loc == "" else row.alt_loc,
+                        alternative_location_id=(
+                            "\x00" if row.alt_loc == "" else row.alt_loc
+                        ),
                         x=row.x_coord,
                         y=row.y_coord,
                         z=row.z_coord,